+++ packages/expo-dev-menu/android/src/debug/java/expo/modules/devmenu/DevMenuFragment.kt
@@ -68,8 +68,7 @@ class DevMenuFragment(
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
 
-    val shouldShowAtLaunch = !DevMenuSessionOverrides.isAutoLaunchDisabled &&
-      (preferences.showsAtLaunch || !preferences.isOnboardingFinished)
+    val shouldShowAtLaunch = preferences.showsAtLaunch || !preferences.isOnboardingFinished
     if (shouldShowAtLaunch) {
       showMenuAtLaunch()
     }